--- core/FPPRG/index.ts
@@ -1,3 +1,0 @@
-export * from './spec'
-export * from './memory'
-export * from './pg'
