+++ db/suggestions.ts
@@ -1,10 +1,10 @@
 import { Table } from '@/spec/sql'
-import * as z from 'zod'
+import { z } from 'zod'
 
 const z_uuid = z.string
 
 export const suggestion = Table.create('suggestion')
-  .field('id', 'uuid', 'primary key default uuid_generate_v4()', z_uuid())
+  .field('id', 'uuid', 'default uuid_generate_v4()', z_uuid(), { primaryKey: true })
   .field('name', 'varchar', 'not null', z.string())
   .field('inputs', 'varchar', 'not null', z.string())
   .field('output', 'varchar', 'not null', z.string())
