+++ spec/sql.ts
@@ -18,9 +18,10 @@ const ObjectCodec = <T>(field_codecs: {[K in keyof T]: Codec<Decoded<T[K]>, Enco
 export type TypedSchema<T = {}> = {
   name: string
   codec: Codec<{[K in keyof T]: Decoded<T[K]>}, {[K in keyof T]: Encoded<T[K]>}>
+  field_pk: {[K in keyof T]: boolean}
+  field_default: {[K in keyof T]: Decoded<T[K]>}
   field_codecs: {[K in keyof T]: Codec<Decoded<T[K]>, Encoded<T[K]>>}
   field_sql: {[K in keyof T]: string}
-  extra_insert_sql?: string,
   schema_up: string
   schema_down: string
 }
@@ -28,11 +29,12 @@ export type TypedSchemaRecord<Schema> = Schema extends TypedSchema<infer T> ? {[
 
 export type TableSchema<T = {}> = {
   name: string
+  field_pk: {[K in keyof T]: boolean}
+  field_default: {[K in keyof T]?: () => Decoded<T[K]>}
   field_codecs: {[K in keyof T]: Codec<Decoded<T[K]>, Encoded<T[K]>>}
   field_sql: {[K in keyof T]: string}
   field_extra_sql: {[K in keyof T]: string}
   extra_sql: string[]
-  extra_insert_sql: string,
 }
 
 /**
@@ -48,11 +50,21 @@ export type TableSchema<T = {}> = {
 export class Table<T = {}> {
   constructor(public t: TableSchema<T>) {}
   static create(name: string) {
-    return new Table({ name, field_codecs: {}, field_sql: {}, field_extra_sql: {}, extra_sql: [], extra_insert_sql: '' })
+    return new Table({
+      name,
+      field_codecs: {},
+      field_pk: {},
+      field_default: {},
+      field_sql: {},
+      field_extra_sql: {},
+      extra_sql: [],
+    })
   }
-  field<S extends string, D, E = D>(name: S, sql: string, extra_sql: string, type: z.ZodType<D> | Codec<D, E>) {
+  field<S extends string, D, E = D>(name: S, sql: string, extra_sql: string, type: z.ZodType<D> | Codec<D, E>, opts?: { primaryKey?: boolean, default?: () => D }) {
     return new Table({
       ...this.t,
+      field_pk: {...this.t.field_pk, [name]: !!(opts||{}).primaryKey },
+      field_default: {...this.t.field_default, [name]: (opts||{}).default },
       field_codecs: { ...this.t.field_codecs, [name]: 'parse' in type ? identityZodCodec(type): type },
       field_sql: { ...this.t.field_sql, [name]: sql },
       field_extra_sql: { ...this.t.field_extra_sql, [name]: extra_sql },
@@ -64,14 +76,8 @@ export class Table<T = {}> {
       extra_sql: [...this.t.extra_sql, extra_sql],
     })
   }
-  extra_insert(extra_insert_sql: string) {
-    return new Table({
-      ...this.t,
-      extra_insert_sql,
-    })
-  }
   build() {
-    const { name, field_codecs, field_sql, extra_insert_sql, field_extra_sql } = this.t
+    const { name, field_codecs, field_pk, field_sql, field_extra_sql } = this.t
     const codec = ObjectCodec(field_codecs)
     const schema_up = [
       `create table ${JSON.stringify(this.t.name)} (`,
@@ -79,13 +85,14 @@ export class Table<T = {}> {
         ...dict.keys(field_sql).map(field =>
           `  ${JSON.stringify(field)} ${field_sql[field]} ${field_extra_sql[field]}`
         ),
+        `primary key (${dict.items(field_pk).filter(({ value }) => value).map(({ key }) => JSON.stringify(key)).join(', ')})`,
         ...this.t.extra_sql.map(sql => `  ${sql}`),
       ].join(',\n'),
       `);`,
     ].join('\n')
     
     const schema_down = `drop table ${JSON.stringify(this.t.name)};`
-    return { codec, name, field_sql, field_codecs, extra_insert_sql, schema_up, schema_down } as TypedSchema<T>
+    return { codec, name, field_sql, field_pk, field_codecs, schema_up, schema_down } as TypedSchema<T>
   }
 }
